+++ rust/bridge/shared/types/src/jni/mod.rs
@@ -286,9 +286,13 @@ impl JniError for SignalProtocolError {
                     jni_args!((addr_name => java.lang.String) -> void),
                 );
             }
-            SignalProtocolError::SessionNotFound(addr) => {
-                let addr_object = protocol_address_to_jobject(env, addr)?;
+            SignalProtocolError::SessionNotFound(SessionNotFound { address, op: _ }) => {
                 let message = new_jstring_from_owned_utf8(env, self.to_string())?;
+                let addr_object = if let Some(address) = address {
+                    protocol_address_to_jobject(env, address)?
+                } else {
+                    JObject::null()
+                };
                 return new_instance(
                     env,
                     ClassName("org.signal.libsignal.protocol.NoSessionException"),